Add Pause Menu
- Simplify system registration and remove unnecessary stages
- Split systems into separate modules
- Add paused state and pause menu
