+++ src/lib.rs
@@ -5,7 +5,6 @@
 use bevy::{
     asset::AssetServerSettings,
     diagnostic::{FrameTimeDiagnosticsPlugin, LogDiagnosticsPlugin},
-    ecs::schedule::ReportExecutionOrderAmbiguities,
     prelude::*,
 };
 use bevy_retro::prelude::*;
@@ -37,7 +36,8 @@ pub fn run() {
         .insert_resource(AssetServerSettings {
             asset_folder: engine_config.asset_path.clone(),
         })
-        .insert_resource(ReportExecutionOrderAmbiguities)
+        // For now, order execution ambiguities are being more annoying than useful
+        // .insert_resource(ReportExecutionOrderAmbiguities)
         // Add engine configuration
         .insert_resource(engine_config.clone())
         // Add the logging config
